refactor(#2299): split resumable compaction streaming into compaction_stream module
Campsite rule (epic #1116): the new stream_partition_body_incremental + CompactionPartitionState/PartitionStreamStep pushed row_decoder/compaction.rs to 898 lines (>800). Extract them to a sibling compaction_stream.rs (328 lines); compaction.rs back to 626. CompactionPolicy + new pub(super) accessors (partition_key/pending_range_start getters+setters) let the split module reuse the exact same decode. No behavior change.
